Checking in code, not sure why UbixOS is no longer able to locate a superblock. I think this may have to do with BLOCK mode but I will do some tests

+++ src/sys/include/pci/hd.h
@@ -173,7 +173,7 @@ struct driveInfo {
   char hdFlags;
   char hdShift;
   long hdMask;
-  long hdMulti;
+  u_int32_t hdMulti;
   long hdPort;
   long hdSize;
   long hdCalc;
